Set method as table name in all AlterTableBase tests
