+++ scylla-cdc-lib/src/test/java/com/scylladb/cdc/lib/AlterUpdateUdtIT.java
@@ -6,6 +6,7 @@ import com.scylladb.cdc.model.worker.ChangeSchema;
 import com.scylladb.cdc.model.worker.RawChange;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInfo;
 
 import java.util.concurrent.TimeUnit;
 
@@ -17,11 +18,6 @@ public class AlterUpdateUdtIT extends AlterTableBase {
   @Override
   public String testKeyspace() {
     return "AlterUpdateUdtIT".toLowerCase();
-  }
-
-  @Override
-  public String testTable() {
-    return "test";
   }
 
   @Override
@@ -138,7 +134,8 @@ public class AlterUpdateUdtIT extends AlterTableBase {
   }
 
   @Test
-  public void alterBeforeNextPageTestBody() {
+  public void alterBeforeNextPageTestBody(TestInfo testInfo) {
+    setTestTableName(testInfo.getTestMethod().get().getName().toLowerCase());
     super.alterBeforeNextPageTestBody();
   }
 }
